clean up and move transapent sync code to a sub module
